requirements.txt
torch>=2.5.0
torchvision>=0.20.0
numpy>=2.0.0
opencv-python>=4.10.0
rembg>=2.0.60
Pillow>=10.0.0
charset-normalizer>=3.3.0
cssselect>=1.2.0
idna>=3.4
lxml>=5.0.0
pyquery>=2.0.0
requests>=2.32.0
tqdm>=4.66.0
urllib3>=2.0.0
dghs-imgutils>=0.8.0
huggingface_hub>=0.25.0,<2.0.0
onnxruntime>=1.20.0
pandas>=2.0.0
